Refactor for improved efficiency and clarity
0. Adds heredocs to all functions.
1. Moves the configuration parsing to the API __init__() to avoid
needing to keep self.faster_whisper_config around.
2. Moves to a single instance of WhisperModel for the API class, as
self.whisper_model, and initializes it during start(). This avoids
needing a separate download step (as the model is downloaded by
WhisperModel's __init__()), reduces the overhead of loading torch and
CUDA models when WhisperModel is initialized, and ensures any potential
memory leaks due to repeated creation of the model are kept to a
minimum.
3. Removes some unneeded imports due to the above refactoring.
4. Adjusts some wording in the README and comments in the flake8 conf.

These changes provide a very significant performance increase for each
transcription call, up to 5 seconds for the first few transcriptions and
up to 1 second for subsequent ones.