Improve transformation handling
Allows transformations to build by iterating over all of them, and
handling each text string with the previously-modified string instead of
casting a list of possible transformations at the start.

For instance this would allow:

it's on -> its on
its on -> lights on

Previously the second would not fire as "its on" was not in the string
at the time the loop began and the list comprehension fired.